✨ Add stopgap raob ingest with rucsoundings kaput
refs #959
